+++ src/include/ovmsg/ovmsgSetPosition.c
@@ -2,7 +2,7 @@
 
 #include "ovmsg_types.h"
 
-void ovmsgSetPosition( OvmsgUID uid, unsigned short x, unsigned short y )
+void ovmsgSetPosition( OvmsgUID uid, unsigned int x, unsigned int y )
 {
 	((struct ovmsg_params *)uid)->x = x;
 	((struct ovmsg_params *)uid)->y = y;
